fix refcount bug in anon_vma code
With the new anon_vma code we take a refcount on the root anon_vma.
However, the root anon_vma does not have a refcount on itself, so
we should not try to do a drop on itself when it is being unlinked.

Just got merged in the -mm kernel.

Fixes bug 602739

Signed-off-by: Rik van Riel <riel@redhat.com>
Tested-by: Dave Young <hidave.darkstar@gmail.com>