rmap: add exclusive page to private anon_vma on swapin
On swapin it is fairly common for a page to be owned exclusively
by one process.  In that case we want to add the page to the
anon_vma of that process's VMA, instead of to the root anon_vma.

This will reduce the amount of rmap searching that the swapout
code needs to do.

Signed-off-by: Rik van Riel <riel@redhat.com>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>