mm: page allocator: Update free page counters after pages are placed on the free list
When allocating a page, the system uses NR_FREE_PAGES counters to determine
in advance if watermarks are being met. This check is made without interrupts
disabled or the zone lock held.

Unfortunately, when pages are being freed in batch, the counters are updated
before the pages are added on the list. During this window, the counters are
misleading as the pages do not exist yet. When under significant pressure
on systems with large numbers of CPUs, it's possible for processes to make
progress even though they should have been stalled. This is particularly
pragmetic if a number of the processes are using GFP_ATOMIC as the min
watermark can be accidentally breached and in extreme cases, the system
can livelock.

This patch updates the counters after the pages have been added to the
list. This makes the allocator more cautious with respect to preserving
the watermarks and mitigates livelock possibilities.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>