mm: vmstat: Actively update vmstat counters in low memory situations
VM stat counters are kept in per-cpu and drained either every stat_interval
seconds or when a threshold is exceeded. This threshold can be quite high
on large systems e.g. 70 on a 16 CPU 8GB system.

NR_FREE_PAGES is a vmstat counter that is depended on for watermark
calculation. The problem is that the counter can be out of sync with
the actual buddy lists allowing the watermarks to be breached. This is
particularly problematic on large systems where there have been reports of
page allocation failures with the "free" counter being significantly higher
than the actual pages on the per-cpu and buddy lists.

Keeping the counters fully in sync would be very expensive. Instead, this
patch has kswapd actively update the vmstat counters while it is awake.
The counters can still be out of sync but within an acceptable margin.

Signed-off-by: Mel Gorman <mel@csn.ul.ie>
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>